requirements.txt
django==5.1.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-filter==24.3
openpyxl==3.1.5
requests==2.32.3

# Добавить для продакшена:
# gunicorn==21.2.0
# psycopg2-binary==2.9.9
# whitenoise==6.6.0
# python-dotenv==1.0.0
# pillow==10.3.0 